Include service IP in service name
I'm adding this to ensure that the service names remain unique. For k8s
we may listen on the same port multiple times and so need to distinguish
by IP. It seems that the config key is just used internally by nerve to
uniquely identify each service. The key isn't stored into ZK so I
*think* it is safe to change the format.