Isolate compose projects from each other
When the different itests run in parallel on the same host, the
containers attach to the same yelp-compose docker network and pollute
the network. This change isolates the itests from each other to prevent
pollution.

Note: There is still pollution if the same itest is run in parallel or
if multiple users run the same itests.